pyproject.toml
[project]
name = "stock-analysis-agent-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-community>=0.3.23",
    "langchain-deepseek>=0.1.3",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "loguru>=0.7.3",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "yfinance>=0.2.59",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
